pyproject.toml
[project]
name = "itsm-multi-agent-support-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Shristi Gautam", email = "shristigautam13@gmail.com" }
]
dependencies = [
    # Add your requirements.txt dependencies here, e.g.:
    "pandas",
    "numpy",
    "python-dotenv",
    "hatchling",
    "pip-tools",
    "pymongo[srv]",
    "certifi",
    "scikit-learn",
    "feast",
    "pyarrow",
    "langgraph",
    "langchain",
    "langchain-core",
    "pydantic",
    "flask"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/itsm_analysis"]

